Migrate log tests
